+++ app/components/BookingButton/wrappers.js
@@ -21,14 +21,14 @@ export const SubButton = styled(MainButton)`
   font-weight: normal;
 
   border: none;
-  background-color: #606060;
+  background-color: ${props => props.theme.colors.darkGrey};
 
   transition: all 200ms linear;
 
   &.sub-btn {
     &--active {
-      background-color: #9fc9eb;
-      color: #191919;
+      background-color: ${props => props.theme.colors.activeSpace};
+      color: ${props => props.theme.colors.black};
     }
   }
 `;